+++ apps/www/src/app/(main)/(docs)/components/tags-input/examples/customized.tsx
@@ -1,0 +1,15 @@
+import { TagsInput } from "@jamsr-ui/react";
+
+export const TagsInputCustomized = () => {
+  return (
+    <TagsInput
+      placeholder="Write your tags here...."
+      label="Tags"
+      slotProps={{
+        chip: {
+          variant: "outlined",
+        },
+      }}
+    />
+  );
+};