+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ui/storybook
 
+## 0.1.5
+
+### Patch Changes
+
+- @jamsr-ui/react@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes