+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/react
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/tags-input@0.1.3
+
 ## 0.1.4
 
 ### Patch Changes